Split reconciler helper and move it to utils
